accessible/tests/mochitest/events/test_textselchange.html

<html>

<head>
  <title>Accessible text selection change events testing</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../text.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>

  <script type="application/javascript">

    async function doTests() {
      // test caret move events and caret offsets

      // synthClick c1_p1: expect caret move, no text selection change
      {
        let p = waitForEvents({
          expected: [[EVENT_TEXT_CARET_MOVED, "c1_p1"]],
          unexpected: [[EVENT_TEXT_SELECTION_CHANGED, "c1_p1"]],
        }, null, false, window);
        getNode("c1_p1").scrollIntoView(true);
        synthesizeMouse(getNode("c1_p1"), 1, 1, {});
        let [caretEvt] = await p;
        caretEvt.QueryInterface(nsIAccessibleCaretMoveEvent);
        is(caretEvt.caretOffset, 0, "Wrong caret offset for c1_p1");
        is(caretEvt.isSelectionCollapsed, true, "Wrong collapsed for c1_p1");
      }

      // shift+down in c1: text selection change event
      {
        let p = waitForEvent(EVENT_TEXT_SELECTION_CHANGED, "c1");
        synthesizeKey("KEY_ArrowDown", { shiftKey: true });
        let selEvt = await p;
        testTextGetSelection("c1", 0, 1, 0);
        selEvt.QueryInterface(nsIAccessibleTextSelectionChangeEvent);
        let selectionRanges = selEvt.selectionRanges;
        let range = selectionRanges.queryElementAt(0, nsIAccessibleTextRange);
        is(
          range.startContainer,
          getAccessible("c1_p1"),
          "correct range start container"
        );
        is(range.startOffset, 0, "correct range start offset");
        is(range.endOffset, 0, "correct range end offset");
        is(
          range.endContainer,
          getAccessible("c1_p2"),
          "correct range end container"
        );
      }

      // shift+down in c1 again
      {
        let p = waitForEvent(EVENT_TEXT_SELECTION_CHANGED, "c1");
        synthesizeKey("KEY_ArrowDown", { shiftKey: true });
        let selEvt = await p;
        testTextGetSelection("c1", 0, 2, 0);
        selEvt.QueryInterface(nsIAccessibleTextSelectionChangeEvent);
        let selectionRanges = selEvt.selectionRanges;
        let range = selectionRanges.queryElementAt(0, nsIAccessibleTextRange);
        is(
          range.startContainer,
          getAccessible("c1_p1"),
          "correct range start container"
        );
        is(range.startOffset, 0, "correct range start offset");
        is(range.endOffset, 9, "correct range end offset");
        is(
          range.endContainer,
          getAccessible("c1_p2"),
          "correct range end container"
        );
      }

      // synthClick ta1: expect caret move, no text selection change
      {
        let p = waitForEvents({
          expected: [[EVENT_TEXT_CARET_MOVED, "ta1"]],
          unexpected: [[EVENT_TEXT_SELECTION_CHANGED, "ta1"]],
        }, null, false, window);
        getNode("ta1").scrollIntoView(true);
        synthesizeMouse(getNode("ta1"), 1, 1, {});
        let [caretEvt] = await p;
        caretEvt.QueryInterface(nsIAccessibleCaretMoveEvent);
        is(caretEvt.caretOffset, 0, "Wrong caret offset for ta1");
        is(caretEvt.isSelectionCollapsed, true, "Wrong collapsed for ta1");
      }

      // shift+right in ta1
      {
        let p = waitForEvent(EVENT_TEXT_SELECTION_CHANGED, "ta1");
        synthesizeKey("KEY_ArrowRight", { shiftKey: true });
        let selEvt = await p;
        testTextGetSelection("ta1", 0, 1, 0);
        selEvt.QueryInterface(nsIAccessibleTextSelectionChangeEvent);
        let selectionRanges = selEvt.selectionRanges;
        let range = selectionRanges.queryElementAt(0, nsIAccessibleTextRange);
        is(
          range.startContainer,
          getAccessible("ta1"),
          "correct range start container"
        );
        is(range.startOffset, 0, "correct range start offset");
        is(range.endOffset, 1, "correct range end offset");
        is(
          range.endContainer,
          getAccessible("ta1"),
          "correct range end container"
        );
      }

      // shift+left in ta1: selection collapses, expect text selection change and caret move
      {
        let p = waitForEvents([
          [EVENT_TEXT_SELECTION_CHANGED, "ta1"],
          [EVENT_TEXT_CARET_MOVED, "ta1"],
        ]);
        synthesizeKey("KEY_ArrowLeft");
        let [, caretEvt] = await p;
        ok(true, "Collapsed selection triggered text selection change event.");
        caretEvt.QueryInterface(nsIAccessibleCaretMoveEvent);
        is(caretEvt.caretOffset, 0, "Wrong caret offset for ta1");
        is(caretEvt.isSelectionCollapsed, true, "Wrong collapsed for ta1");
      }

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTests);
  </script>
</head>

<body>

  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=762934"
     title="Text selection change event has a wrong target when selection is spanned through several objects">
    Bug 762934
  </a>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=956032"
     title="Text selection change event missed when selected text becomes unselected">
    Bug 956032
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <div id="c1" contentEditable="true">
    <p id="c1_p1">paragraph</p>
    <p id="c1_p2">paragraph</p>
  </div>

  <textarea id="ta1">Hello world</textarea>

  <div id="eventdump"></div>
</body>
</html>
